+++ CHANGELOG.md
@@ -8,6 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### v3.9.x - [Unreleased]
 
+#### Changed
+
+- Extra Nginx directives moved from nginx.conf to conf.d/tweaks.conf
+
+#### Fixed
+
+- MySQLTuner installation
+- `wo stack remove/purge --all`
+- variable substitution in install script
+
 ### v3.9.8 - 2019-08-16
 
 #### Added